requirements.txt
Flask
transformers
torch
sentence-transformers
scikit-learn
requests
python-dotenv